Finish migrating winsock2 to 0.3
