+++ build.rs
@@ -53,6 +53,7 @@ const DATA: &'static [(&'static str, &'static [&'static str], &'static [&'static
     ("windef", &["minwindef", "winnt"], &[]),
     ("windowsx", &["minwindef"], &[]),
     ("winerror", &["minwindef"], &[]),
+    ("ws2def", &["guiddef", "inaddr", "minwindef", "vcruntime", "winnt"], &[]),
     ("wtypes", &["guiddef", "minwindef", "ntdef", "wtypesbase"], &[]),
     ("wtypesbase", &["minwindef", "rpcndr", "winnt"], &[]),
     // ucrt
@@ -106,7 +107,7 @@ const DATA: &'static [(&'static str, &'static [&'static str], &'static [&'static
     ("gl-gl", &[], &[]),
     ("handleapi", &["minwindef", "winnt"], &["kernel32"]),
     ("heapapi", &["basetsd", "minwindef"], &[]),
-    ("http", &["guiddef", "minwindef", "sspi", "winnt", "winsock2"], &[]),
+    ("http", &["guiddef", "minwindef", "sspi", "winnt", "ws2def"], &[]),
     ("imm", &[], &[]),
     ("libloaderapi", &["basetsd", "minwindef", "winnt"], &["kernel32", "user32"]),
     ("lmaccess", &["lmcons", "minwindef", "winnt"], &[]),
@@ -184,7 +185,7 @@ const DATA: &'static [(&'static str, &'static [&'static str], &'static [&'static
     ("winreg", &["basetsd", "minwinbase", "minwindef", "winnt"], &["advapi32"]),
     ("winscard", &["basetsd", "guiddef", "minwindef", "windef", "winnt"], &[]),
     ("winsmcrd", &["minwindef", "winioctl"], &[]),
-    ("winsock2", &["guiddef", "inaddr", "minwindef", "vcruntime", "winnt"], &[]),
+    ("winsock2", &["basetsd", "guiddef", "minwinbase", "minwindef", "qos", "windef", "winerror", "winnt", "ws2def", "wtypesbase"], &[]),
     ("winspool", &["minwindef", "wingdi", "winnt"], &[]),
     ("winsvc", &["minwindef", "winnt"], &["advapi32"]),
     ("winusb", &["minwindef", "winnt"], &[]),
