+++ src/um/http.rs
@@ -8,11 +8,11 @@
 use shared::guiddef::GUID;
 use shared::minwindef::{DWORD, PUCHAR, UCHAR, ULONG, USHORT};
 use shared::sspi::SECURITY_STATUS;
+use shared::ws2def::{PSOCKADDR, SOCKADDR_STORAGE};
 use um::winnt::{
     ANYSIZE_ARRAY, BOOLEAN, HANDLE, PCHAR, PCSTR, PCWSTR, PSECURITY_DESCRIPTOR, PVOID, PWCHAR,
     PWSTR, ULARGE_INTEGER, ULONGLONG,
 };
-use um::winsock2::{PSOCKADDR, SOCKADDR_STORAGE};
 pub const HTTP_INITIALIZE_SERVER: ULONG = 0x00000001;
 pub const HTTP_INITIALIZE_CONFIG: ULONG = 0x00000002;
 pub const HTTP_DEMAND_CBT: ULONG = 0x00000004;
